Move  Instructions Page Card 2 into  InstructionsPage folder
