+++ src/components/InstructionsPage/Card2/instruction.css
